[DepthMapEntity] Add try/catch when reading the depth and sim maps
The depth and sim maps are identified by taking the filename of the loaded
file (may it be the depth map or the sim map) and finding its counterpart
by replacing "depthMap" by "simMap" (or "simMap" by "depthMap") in the
filename. Whether the files exist is not checked until we attempt to read
the maps.

image::readImage throws a std::runtime_error if the file does not exist or
if some other issues occur while reading it. As both the depth map and sim
map are expected, if one is missing, we will still attempt to read it and
an exception will be thrown without being handled, causing the application
to crash.

This commit tries and catches std::runtime_error exceptions when reading
the images. If the depth map does not exist, the loading process is
interrupted and the DepthMapEntity status is updated to "Error". If the
sim map does not exist, we keep on going with the whole process: the sim
map will be deemed as invalid, and only the depth map will be displayed.

This prevents crashes from happening in any case where there might be a
depth map without a corresponding sim map.

Some log levels have also been increased from "debug" to "warning" when
they indicate issues.