[qmlSfmData] Only disable cameras when resections display is enabled
Keep track of whether the display of resections is enabled or not:
if it is, then the cameras can be shown or hidden based on their resection
ID. Otherwise, they should all be explicitly displayed.

This commit fixes an issue where, if two SfMs with a visibly different
number of camera reconstruction groups were displayed at the same time,
selecting the one with the fewer number of groups was automatically hiding
some of the cameras of the one that had more, independently from the
slider's status.