[qtAliceVision] FloatImageViewer: Add a global status
Instead of only having a loading status that oscillates between the "loading"
and "not loading" state, add a global status for the viewer that can describe
more states.

The global status is updated depending on the loading status and on the error
flag of the `ResponseData` object.

The loading status itself is kept as it allows to distinguish several
loading use cases but is not exposed to the QML anymore, as the global status
covers all the relevant information.