[qt6] PanoramaViewer: Fix hovering effects
When images in the PanoramaViewer are hovered, their gamma is increased
to highlight them and make it obvious to the user which image is currently
being seen.

This commit fixes an issue that occurred when the mouse was moving away
from the image it was hovering: the added gamma was never removed, meaning
that, by hovering one image enough times, said image would become totally
white. The highlighting effect now correctly increases and decreases the
hovered image's gamma level, including when the user clicks to edit the
panorama.