[qmlSfmData] SfmDataEntity: Handle fixed point size in the shader
The size of the points in the cloud is now fully handled in the
dedicated cloud shader. Prior to this commit, the size of the points
in the shader was always determined relatively to the position of the
camera. When the `fixedPointSize` option is set, it is now set
independently from the camera's position.

The `fixedPointSize` variable is added with its corresponding signal and
slot.